backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Google APIs for Gmail integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# AI/ML
openai==1.3.5
langchain==0.0.350
langchain-openai==0.0.2

# Background tasks
redis==5.0.1
rq==1.15.1

# Data processing
pydantic==2.5.0
pandas==2.1.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
